refactor - Renamed the color convention to legend.
