+++ charts/src/main/java/dev/anirban/charts/circular/CircularChartInterface.kt
@@ -5,7 +5,7 @@ import androidx.compose.ui.Modifier
 import androidx.compose.ui.graphics.drawscope.DrawScope
 import dev.anirban.charts.circular.center.CircularCenterInterface
 import dev.anirban.charts.circular.decoration.CircularDecoration
-import dev.anirban.charts.circular.colorconvention.CircularColorConventionInterface
+import dev.anirban.charts.circular.legend.CircularLegendInterface
 import dev.anirban.charts.circular.data.CircularDataInterface
 import dev.anirban.charts.circular.foreground.CircularForegroundInterface
 
@@ -41,7 +41,7 @@ interface CircularChartInterface {
     /**
      * This implementation draws the Color Conventions for the Chart
      */
-    val circularColorConvention: CircularColorConventionInterface
+    val circularColorConvention: CircularLegendInterface
 
     /**
      * This function draws the center of the Chart
