+++ charts/src/main/java/dev/anirban/charts/circular/legend/CircularLegendInterface.kt
@@ -1,4 +1,4 @@
-package dev.anirban.charts.circular.colorconvention
+package dev.anirban.charts.circular.legend
 
 import androidx.compose.runtime.Composable
 import dev.anirban.charts.circular.decoration.CircularDecoration
@@ -8,13 +8,13 @@ import dev.anirban.charts.circular.data.CircularDataInterface
  * This implementation shall be implemented by all the classes which are making color convention
  * implementation
  *
- * Implementations for this interface are :- [CircularNoColorConvention],
- * [CircularGridColorConvention],[CircularListColorConvention],[CircularTargetColorConvention]
+ * Implementations for this interface are :- [CircularNoLegend],
+ * [CircularGridLegend],[CircularListLegend],[CircularTargetLegend]
 
  *
  * @property DrawColorConventions THis function draws the desired color Convention
  */
-interface CircularColorConventionInterface {
+interface CircularLegendInterface {
 
     @Composable
     fun DrawColorConventions(
